+++ pkg/services/object/metrics.go
@@ -4,7 +4,9 @@ import (
 	"context"
 	"time"
 
+	"github.com/mr-tron/base58"
 	"github.com/nspcc-dev/neofs-api-go/v2/object"
+	"github.com/nspcc-dev/neofs-api-go/v2/refs"
 	"github.com/nspcc-dev/neofs-node/pkg/services/util"
 )
 
@@ -18,22 +20,24 @@ type (
 		util.ServerStream
 		stream  GetObjectStream
 		metrics MetricRegister
+		cid     string
 	}
 
 	putStreamMetric struct {
 		stream  PutObjectStream
 		metrics MetricRegister
 		start   time.Time
+		cid     string
 	}
 
 	MetricRegister interface {
-		IncGetReqCounter(success bool)
-		IncPutReqCounter(success bool)
-		IncHeadReqCounter(success bool)
-		IncSearchReqCounter(success bool)
-		IncDeleteReqCounter(success bool)
-		IncRangeReqCounter(success bool)
-		IncRangeHashReqCounter(success bool)
+		IncGetReqCounter(cnrID string, success bool)
+		IncPutReqCounter(cnrID string, success bool)
+		IncHeadReqCounter(cnrID string, success bool)
+		IncSearchReqCounter(cnrID string, success bool)
+		IncDeleteReqCounter(cnrID string, success bool)
+		IncRangeReqCounter(cnrID string, success bool)
+		IncRangeHashReqCounter(cnrID string, success bool)
 
 		AddGetReqDuration(time.Duration)
 		AddPutReqDuration(time.Duration)
@@ -58,7 +62,8 @@ func NewMetricCollector(next ServiceServer, register MetricRegister) *MetricColl
 func (m MetricCollector) Get(req *object.GetRequest, stream GetObjectStream) (err error) {
 	t := time.Now()
 	defer func() {
-		m.metrics.IncGetReqCounter(err == nil)
+		cnr := getCIDString(req.GetBody().GetAddress().GetContainerID())
+		m.metrics.IncGetReqCounter(cnr, err == nil)
 		m.metrics.AddGetReqDuration(time.Since(t))
 	}()
 
@@ -90,7 +95,8 @@ func (m MetricCollector) Head(ctx context.Context, request *object.HeadRequest) 
 
 	res, err := m.next.Head(ctx, request)
 
-	m.metrics.IncHeadReqCounter(err == nil)
+	cnr := getCIDString(request.GetBody().GetAddress().GetContainerID())
+	m.metrics.IncHeadReqCounter(cnr, err == nil)
 	m.metrics.AddHeadReqDuration(time.Since(t))
 
 	return res, err
@@ -101,7 +107,8 @@ func (m MetricCollector) Search(req *object.SearchRequest, stream SearchStream) 
 
 	err := m.next.Search(req, stream)
 
-	m.metrics.IncSearchReqCounter(err == nil)
+	cnr := getCIDString(req.GetBody().GetContainerID())
+	m.metrics.IncSearchReqCounter(cnr, err == nil)
 	m.metrics.AddSearchReqDuration(time.Since(t))
 
 	return err
@@ -112,7 +119,8 @@ func (m MetricCollector) Delete(ctx context.Context, request *object.DeleteReque
 
 	res, err := m.next.Delete(ctx, request)
 
-	m.metrics.IncDeleteReqCounter(err == nil)
+	cnr := getCIDString(request.GetBody().GetAddress().GetContainerID())
+	m.metrics.IncDeleteReqCounter(cnr, err == nil)
 	m.metrics.AddDeleteReqDuration(time.Since(t))
 
 	return res, err
@@ -123,7 +131,8 @@ func (m MetricCollector) GetRange(req *object.GetRangeRequest, stream GetObjectR
 
 	err := m.next.GetRange(req, stream)
 
-	m.metrics.IncRangeReqCounter(err == nil)
+	cnr := getCIDString(req.GetBody().GetAddress().GetContainerID())
+	m.metrics.IncRangeReqCounter(cnr, err == nil)
 	m.metrics.AddRangeReqDuration(time.Since(t))
 
 	return err
@@ -134,7 +143,8 @@ func (m MetricCollector) GetRangeHash(ctx context.Context, request *object.GetRa
 
 	res, err := m.next.GetRangeHash(ctx, request)
 
-	m.metrics.IncRangeHashReqCounter(err == nil)
+	cnr := getCIDString(request.GetBody().GetAddress().GetContainerID())
+	m.metrics.IncRangeHashReqCounter(cnr, err == nil)
 	m.metrics.AddRangeHashReqDuration(time.Since(t))
 
 	return res, err
@@ -161,8 +171,13 @@ func (s putStreamMetric) Send(req *object.PutRequest) error {
 func (s putStreamMetric) CloseAndRecv() (*object.PutResponse, error) {
 	res, err := s.stream.CloseAndRecv()
 
-	s.metrics.IncPutReqCounter(err == nil)
+	s.metrics.IncPutReqCounter(s.cid, err == nil)
 	s.metrics.AddPutReqDuration(time.Since(s.start))
 
 	return res, err
+}
+
+func getCIDString(cnrV2 *refs.ContainerID) string {
+	// Must be equal to `EncodeToString` for container ID from SDK.
+	return base58.Encode(cnrV2.GetValue())
 }